run the term frequency matrix on the list of urls provided as parameter.

The function download the pdf, create the corpus and generate the term frequency matrix, to avoid storage limitation the function deleate original pdf files

scrap.ccdr.tfs(
  urls = url_links,
  lexicon,
  extract_number = 1,
  ENGINE = pdf_text,
  delete_pdfs = T,
  rm_short_docs = F,
  min_words = 100,
  parrallel = T,
  loc_temp = NULL
)

Arguments

urls
a dataframe containing two columns: name_file and pdf that respectively provid the name that will be given to the file downloaded and pdf that contains the url of the pdf

lexicon
one of the element of the list provided by lexicon()

extract_number
a number that will be ued as a suffix for the name of the corpus and the name of the tf matrix

ENGINE
function to read pdf into environment. Either pdf_text or pdf_ocr_text.

delete_pdfs
T/F, set to T it will delete the folder containing the original pdf, usefull option when the number of pdf is very large and the size of the folder start to be very large

rm_short_docs
T/F T if you want to remove the documents under a certain number of words

min_words
the minimum word in the document necessary to perform the text mining

parrallel
if T then will use mclapply from the package parrallel

loc_temp
path to the folder containing the files, corpus and tf

Value

create forlders where pdfs, corpus and tf-idf matrix are stored

Author

Manuel Betin